pyproject.toml
[tool.poetry]
name = "hw20"
version = "0.1.0"
description = ""
authors = ["max"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
django = "4.2"
ipython = "^8.24.0"
djangorestframework = "^3.15.1"
pillow = "^10.3.0"
psycopg2 = "^2.9.9"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
